requirements.txt
openai
langchain
langgraph
python-dotenv
requests
transformers
tiktoken
pydantic
langchain_community
torch
langchain-openai
duckduckgo-search
wikipedia
langchain_chroma
pypdf
langchain-mcp-adapters
mcp
huggingface_hub 
pillow
fastapi
python-telegram-bot
streamlit
